# improves Stakes::activate_epoch performance